commit! no crush by deleting not existing task

+++ app/main.py
@@ -73,11 +73,14 @@ async def todo_edit(
     return RedirectResponse(url=app.url_path_for("home"), status_code=status.HTTP_303_SEE_OTHER)
 
 
-@app.get("/delete/{todo_id}")
-async def todo_delete(request: Request, todo_id: int, database: Session = Depends(get_db)):
+@app.get("/delete/{todo_id}", status_code=status.HTTP_200_OK)
+async def todo_delete(response: Response, request: Request, todo_id: int, database: Session = Depends(get_db)):
     """Delete todo
     """
     todo = database.query(models.Todo).filter(models.Todo.id == todo_id).first()
+    if todo is None:
+        response.status_code = status.HTTP_404_NOT_FOUND
+        return RedirectResponse(url=app.url_path_for("home"), status_code=status.HTTP_303_SEE_OTHER)
     logger.info(f"Deleting todo: {todo}")
     database.delete(todo)
     database.commit()
